Fall back to key if no source language value
getSourceLanguageTranslationsWith() only takes strings from the language
files, not from the application source code.

If you call auto-translate with a language parameter, it'll add missing
keys to the target language, but not to the source language files.

We work around this by falling back to the key if no value is set in the
source language.